requirements.txt
anthropic>=0.39.0
fastapi>=0.115.0
uvicorn>=0.32.0
pyyaml>=6.0
reportlab>=4.1.0
pdfplumber>=0.11.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
